refactor(processor): extract shared region measurement runner
  - Move region validation, pre-roll seek, and filter graph setup into a
    shared helper.
  - Reuse it for band and output region analysis to remove duplicated code.
